requirements.txt
opencv-python
pytesseract
pyautogui
lxml
xmltodict
keyboard
pygetwindow
pillow
